chore(submitentityclaim):  separate page
