+++ src/common/components/ControlPanel/ControlPanel.styles.ts
