+++ src/common/components/ControlPanel/schema/Cell.schema.json
