+++ src/common/components/ControlPanel/schema/Project.schema.json
