+++ src/components/Routes.tsx
@@ -3,6 +3,7 @@ import { Route, Switch } from 'react-router-dom'
 import { ProjectContainerConnected } from './project/ProjectContainer'
 import { EntitiesContainerConnected } from '../modules/Entities/Entities.container'
 import { EntitiesSelectConnected } from '../modules/Entities/Entities.container.select'
+import { SubmitEntityClaimConnected } from '../modules/SubmitEntityClaim/SubmitEntityClaim.container'
 import { ProjectCreateConnected } from './project/ProjectCreate'
 import { contentType } from '../types/models'
 import { ProjectForm } from './project/ProjectForm'
@@ -66,6 +67,7 @@ export const Routes: React.SFC<{}> = props => {
           )}
         />
         <Route
+          exact
           path="/projects/:projectDID/overview"
           render={(): JSX.Element => (
             <ProjectContainerConnected
@@ -73,6 +75,11 @@ export const Routes: React.SFC<{}> = props => {
               contentType={contentType.overview}
             />
           )}
+        />
+        <Route
+          exact
+          path={`/projects/:projectDID/overview/action/new_claim`}
+          component={SubmitEntityClaimConnected}
         />
         {/* Old claims related screens - remove when new claims is ready */}
         {/*        <Route
