+++ src/modules/SubmitEntityClaim/SubmitEntityClaim.container.tsx
@@ -6,13 +6,18 @@ import Question from './components/Question/Question'
 import { Container } from './SubmitEntityClaim.container.styles'
 import { FormControl } from '../../common/components/JsonForm/types'
 import * as submitEntityClaimSelectors from './SubmitEntityClaim.selectors'
+import * as accountSelectors from '../Account/Account.selectors'
 import {
   goToNextQuestion,
   goToPreviousQuestion,
 } from './SubmitEntityClaim.actions'
-import { ActionWrapper } from '../../components/project/ControlPanel/Actions/Actions.styles'
+import ControlPanel from '../../common/components/ControlPanel/ControlPanel'
+import CellControlPanelSchema from '../../common/components/ControlPanel/schema/Cell.schema.json'
+
+// TODO - hookup redux for project data
 
 interface Props {
+  userDid: string
   currentQuestion: FormControl
   currentQuestionNo: number
   questionCount: number
@@ -40,6 +45,7 @@ class SubmitEntityClaim extends React.Component<Props, State> {
 
   render(): JSX.Element {
     const {
+      userDid,
       currentQuestion,
       currentQuestionNo,
       questionCount,
@@ -47,21 +53,36 @@ class SubmitEntityClaim extends React.Component<Props, State> {
       handleNextClick,
     } = this.props
     return (
-      <ActionWrapper className="open">
-        <Container>
-          {this.state.showInstructions ? (
-            <Instructions toggleInstructions={this.handleToggleInstructions} />
-          ) : (
-            <Question
-              handlePreviousClick={handlePreviousClick}
-              handleNextClick={handleNextClick}
-              question={currentQuestion}
-              currentQuestionNo={currentQuestionNo}
-              questionCount={questionCount}
-            />
-          )}
-        </Container>
-      </ActionWrapper>
+      <div className="container-fluid">
+        <div className="container">
+          <div className="row">
+            <div className="col-lg-8">
+              <Container>
+                {this.state.showInstructions ? (
+                  <Instructions
+                    toggleInstructions={this.handleToggleInstructions}
+                  />
+                ) : (
+                  <Question
+                    handlePreviousClick={handlePreviousClick}
+                    handleNextClick={handleNextClick}
+                    question={currentQuestion}
+                    currentQuestionNo={currentQuestionNo}
+                    questionCount={questionCount}
+                  />
+                )}
+              </Container>
+            </div>
+            <div className="col-lg-4">
+              <ControlPanel
+                schema={CellControlPanelSchema}
+                entityDid={'123'}
+                userDid={userDid}
+              />
+            </div>
+          </div>
+        </div>
+      </div>
     )
   }
 }
@@ -70,6 +91,7 @@ const mapStateToProps = (state: RootState): Record<string, any> => ({
   currentQuestion: submitEntityClaimSelectors.selectCurrentQuestion(state),
   currentQuestionNo: submitEntityClaimSelectors.selectCurrentQuestionNo(state),
   questionCount: submitEntityClaimSelectors.selectQuestionCount(state),
+  userDid: accountSelectors.selectUserDid(state),
 })
 
 const mapDispatchToProps = (dispatch: Dispatch<any>): any => ({
@@ -77,4 +99,7 @@ const mapDispatchToProps = (dispatch: Dispatch<any>): any => ({
   handleNextClick: (): void => dispatch(goToNextQuestion()),
 })
 
-export default connect(mapStateToProps, mapDispatchToProps)(SubmitEntityClaim)
+export const SubmitEntityClaimConnected = connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(SubmitEntityClaim)
